+++ tests/editor_preview_backdrop_smoke.gd
@@ -3,13 +3,13 @@ extends SceneTree
 const PREVIEW_SCENES := [
 	"res://gameplay/expeditions/expedition_altar.tscn",
 	"res://environment/props/sanctuary/divine_fountain.tscn",
-	"res://environment/props/sanctuary/mushroom_dwelling.tscn",
-	"res://environment/props/sanctuary/merchant_hall.tscn",
-	"res://environment/props/sanctuary/weapon_stall.tscn",
+	"res://environment/props/sanctuary/skillkeeper_lodge.tscn",
+	"res://environment/props/sanctuary/armskeeper_workshop.tscn",
+	"res://environment/props/sanctuary/armskeeper_cart.tscn",
 	"res://environment/props/sanctuary/sanctuary_tree_broad.tscn",
 	"res://environment/props/sanctuary/sanctuary_tree_tall.tscn",
 	"res://entities/npcs/skillkeeper/skillkeeper.tscn",
-	"res://entities/npcs/weapon_merchant/weapon_merchant.tscn",
+	"res://entities/npcs/armskeeper/armskeeper.tscn",
 ]